+++ docs/changelog.rst
@@ -6,6 +6,11 @@ Changelog
 * :feature:`-` rotki will now properly decode the swaps done via Uniswap V3 on other supported chains.
 * :feature:`5978` rotki will now properly decode the swaps done via the 0x protocol.
 * :feature:`-` The claim event of the Degen airdrop 2 will be correctly decoded.
+
+* :release:`1.32.2 <2024-03-15>`
+* :bug:`7531` rotki will now prioritize manual prices above other price oracles and calculations.
+* :bug:`-` rotki will display NFT prices correctly across all asset types.
+* :bug:`7572` Fix issue where coinbase data query could result in an error popup.
 * :bug:`7611` USDT balances and trades in bitcoin.de should now work properly.
 * :bug:`-` Sending money to SEPA or bridging with the BurnFrom monerium signature will now be properly decoded by rotki.
 * :bug:`-` Fix issue where long label in manual balances breaks the alignment of chain names.